rag/requirement.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-ollama
chromadb
sentence-transformers
rank_bm25
pymupdf
streamlit
fastapi
uvicorn
asyncio
python-dotenv
